requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
pyyaml==6.0.1
numpy==1.26.2
scikit-learn==1.3.2
rank-bm25==0.2.2
python-multipart==0.0.6
